Adjusted concurrent event tests
Under highly concurrent usage in publishing it's 18129ms vs 342ms so 53x
faster. And Glader.Essentials.EventBus actually supports even better
concurrency when not broadcasting the same event type from multiple
threads too. So this is actually the worst cast scenario for
Glader.Essentials.EventBus and the only case scenario for Redbus as it
functions the same regardless of event type.